@@ -2,7 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
 
-ReactDOM.render(
-  <App />,
+const root = ReactDOM.createRoot(
   document.getElementById('root')
 );
+
+root.render(<App/>);
